Store test patches as fp16 to halve CPU RAM during OSE inference
Float32 accumulation across 337 batches (29×680×1440 each) reached ~38 GB,
triggering the Linux OOM killer at batch 329. Storing as half-precision
reduces peak RAM to ~19 GB.

+++ src/lit_flow_matching_osse_sla_oceanfm.py
@@ -274,11 +274,12 @@ class LitFlowMatchingOSSE_SLA_OceanFM(pl.LightningModule):
             for k in range(self.n_ensemble):
                 x_0 = self._get_source(batch)
                 x_member = self._sample_ode(condition, x_0, use_ema=True)
-                self.test_data_members[k].append(x_member.detach().cpu().unsqueeze(1) * s + m)
+                # store as fp16 to halve CPU RAM usage (~19 GB vs ~38 GB for 337 batches)
+                self.test_data_members[k].append((x_member.detach().cpu().unsqueeze(1) * s + m).half())
         else:
             x_0 = self._get_source(batch)
             x_refined = self._sample_ode(condition, x_0, use_ema=True)
-            self.test_data.append(x_refined.detach().cpu().unsqueeze(1) * s + m)
+            self.test_data.append((x_refined.detach().cpu().unsqueeze(1) * s + m).half())
 
     # ------------------------------------------------------------------
     # Optimizer
